create CommonUi dll and adjust BruteSharkCli accordingly
