+++ BruteShark/CommonUi/CommonUi.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\BruteForce\BruteForce.csproj" />
+    <ProjectReference Include="..\PcapAnalyzer\PcapAnalyzer.csproj" />
+    <ProjectReference Include="..\PcapProcessor\PcapProcessor.csproj" />
+  </ItemGroup>
+
+</Project>
